[ANI-68-ani-path-detail] refactor: domain => api  파일명을 변경합니다.
