+++ apps/web/src/views/path/api/get-path-introduce.ts
@@ -1,6 +1,6 @@
 import { HTTP_HEADERS, HTTP_METHODS } from '@/shared/config/constants/http'
 import type { BasePartType } from '@/shared/types/path'
-import type { PathDetailType, PathIntroduceType } from '../model/domain'
+import type { PathDetailType, PathIntroduceType } from '../model/apis'
 
 type Data = {
   path: PathDetailType
